[wmco] Make operator manage windows-instances
The "windows-instances" ConfigMap is the single source of truth for BYOH
instances. Given that make the operator manage it. This means that it
will be created on operator bootup if not present. If the admin deletes
it, the operation is viewed as a call to deconfigure all BYOH Windows
instances. The ConfigMap will then be re-created with an empty data
section.